Refactor DcpExecutor into core and object creators
# Conflicts:
#	src/Aspire.Hosting/Dcp/DcpExecutor.cs
