+++ src/Aspire.Hosting/Dcp/DcpModelUtilities.cs
@@ -1,0 +1,114 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System.Diagnostics;
+using System.Globalization;
+using System.Net;
+using Aspire.Hosting.ApplicationModel;
+using Aspire.Hosting.Dcp.Model;
+
+namespace Aspire.Hosting.Dcp;
+
+/// <summary>
+/// Static utility methods for working with the DCP model.
+/// </summary>
+internal static class DcpModelUtilities
+{
+    /// <summary>
+    /// Examines the Aspire resource annotations and adds equivalent ServiceProducerAnnotations to the corresponding DCP resource.
+    /// </summary>
+    internal static void AddServicesProducedInfo<TDcpResource>(
+        RenderedModelResource<TDcpResource> appResource,
+        IEnumerable<IAppResource> appResources)
+        where TDcpResource : CustomResource, IKubernetesStaticMetadata
+    {
+        var modelResource = appResource.ModelResource;
+        var modelResourceName = modelResource.Name ?? "(unknown)";
+
+        var servicesProduced = appResources.OfType<ServiceWithModelResource>().Where(r => r.ModelResource == modelResource);
+        foreach (var sp in servicesProduced)
+        {
+            var ea = sp.EndpointAnnotation;
+
+            if (modelResource.IsContainer())
+            {
+                if (ea.TargetPort is null)
+                {
+                    throw new InvalidOperationException($"The endpoint '{ea.Name}' for container resource '{modelResourceName}' must specify the {nameof(EndpointAnnotation.TargetPort)} value");
+                }
+            }
+            else if (!ea.IsProxied)
+            {
+                if (HasMultipleReplicas(appResource.DcpResource))
+                {
+                    throw new InvalidOperationException($"Resource '{modelResourceName}' uses multiple replicas and a proxy-less endpoint '{ea.Name}'. These features do not work together.");
+                }
+
+                if (ea.Port is int && ea.Port != ea.TargetPort)
+                {
+                    throw new InvalidOperationException($"The endpoint '{ea.Name}' for resource '{modelResourceName}' is not using a proxy, and it has a value of {nameof(EndpointAnnotation.Port)} property that is different from the value of {nameof(EndpointAnnotation.TargetPort)} property. For proxy-less endpoints they must match.");
+                }
+            }
+            else
+            {
+                Debug.Assert(ea.IsProxied);
+
+                if (ea.TargetPort is int && ea.Port is int && ea.TargetPort == ea.Port)
+                {
+                    throw new InvalidOperationException(
+                        $"The endpoint '{ea.Name}' for resource '{modelResourceName}' requested a proxy ({nameof(ea.IsProxied)} is true). Non-container resources cannot be proxied when both {nameof(ea.TargetPort)} and {nameof(ea.Port)} are specified with the same value.");
+                }
+
+                if (HasMultipleReplicas(appResource.DcpResource) && ea.TargetPort is int)
+                {
+                    throw new InvalidOperationException(
+                        $"Resource '{modelResourceName}' can have multiple replicas, and it uses endpoint '{ea.Name}' that has {nameof(ea.TargetPort)} property set. Each replica must have a unique port; setting {nameof(ea.TargetPort)} is not allowed.");
+                }
+            }
+
+            var spAnn = new ServiceProducerAnnotation(sp.Service.Metadata.Name);
+            (spAnn.Address, _) = NormalizeTargetHost(ea.TargetHost);
+            spAnn.Port = ea.TargetPort;
+            appResource.DcpResource.AnnotateAsObjectList(CustomResource.ServiceProducerAnnotation, spAnn);
+            appResource.ServicesProduced.Add(sp);
+        }
+
+        static bool HasMultipleReplicas(CustomResource resource)
+        {
+            if (resource is Executable exe && exe.Metadata.Annotations.TryGetValue(CustomResource.ResourceReplicaCount, out var value) && int.TryParse(value, CultureInfo.InvariantCulture, out var replicas) && replicas > 1)
+            {
+                return true;
+            }
+            return false;
+        }
+    }
+
+    /// <summary>
+    /// Normalize the target host to a tuple of (address, binding mode) to a single valid address for
+    /// service discovery purposes. A user may have configured an endpoint target host that isn't itself
+    /// a valid IP address or hostname that can be resolved by other services or clients. For example,
+    /// 0.0.0.0 is considered to mean that the service should bind to all IPv4 addresses. When the target
+    /// host indicates that the service should bind to all IPv4 or IPv6 addresses, we instead return
+    /// "localhost" as the address as that is a valid address for the .NET dev certificate. The binding mode
+    /// is metadata that indicates whether an endpoint is bound to a single address or some set of multiple
+    /// addresses on the system.
+    /// </summary>
+    /// <param name="targetHost">The target host from an EndpointAnnotation</param>
+    /// <returns>A tuple of (address, binding mode).</returns>
+    internal static (string, EndpointBindingMode) NormalizeTargetHost(string targetHost)
+    {
+        return targetHost switch
+        {
+            null or "" => (KnownHostNames.Localhost, EndpointBindingMode.SingleAddress), // Default is localhost
+            var s when EndpointHostHelpers.IsLocalhostOrLocalhostTld(s) => (KnownHostNames.Localhost, EndpointBindingMode.SingleAddress), // Explicitly set to localhost or .localhost subdomain
+
+            var s when IPAddress.TryParse(s, out var ipAddress) => ipAddress switch // The host is an IP address
+            {
+                var ip when IPAddress.Any.Equals(ip) => (KnownHostNames.Localhost, EndpointBindingMode.IPv4AnyAddresses), // 0.0.0.0 (IPv4 all addresses)
+                var ip when IPAddress.IPv6Any.Equals(ip) => (KnownHostNames.Localhost, EndpointBindingMode.IPv6AnyAddresses), // :: (IPv6 all addresses)
+                _ => (s, EndpointBindingMode.SingleAddress), // Any other IP address is returned as-is as that will be the only address the service is bound to
+            },
+            _ => (KnownHostNames.Localhost, EndpointBindingMode.DualStackAnyAddresses), // Any other target host is treated as binding to all IPv4 AND IPv6 addresses
+        };
+    }
+}
