+++ src/Aspire.Hosting/Dcp/DcpResourceState.cs
@@ -7,7 +7,7 @@ using Aspire.Hosting.Dcp.Model;
 
 namespace Aspire.Hosting.Dcp;
 
-internal sealed class DcpResourceState(IDictionary<string, IResource> applicationModel, IEnumerable<AppResource> appResources)
+internal sealed class DcpResourceState(IDictionary<string, IResource> applicationModel, IEnumerable<IAppResource> appResources)
 {
     public readonly ConcurrentDictionary<string, Container> ContainersMap = [];
     public readonly ConcurrentDictionary<string, Executable> ExecutablesMap = [];
@@ -17,5 +17,5 @@ internal sealed class DcpResourceState(IDictionary<string, IResource> applicatio
     public readonly ConcurrentDictionary<(string, string), List<string>> ResourceAssociatedServicesMap = [];
 
     public IDictionary<string, IResource> ApplicationModel { get; } = applicationModel;
-    public IEnumerable<AppResource> AppResources { get; } = appResources;
+    public IEnumerable<IAppResource> AppResources { get; } = appResources;
 }
